2: refactor program structures and add bump in player state
