+++ programs/blockrunners/src/errors.rs
@@ -1,0 +1,7 @@
+use anchor_lang::error_code;
+
+#[error_code]
+pub enum BlockRunnersError {
+    #[msg("Unknown Error")]
+    UnknownError,
+}
