+++ programs/blockrunners/src/instructions/mod.rs
@@ -1,0 +1,2 @@
+pub mod join_game;
+pub use join_game::*;
