+++ programs/blockrunners/src/lib.rs
@@ -2,6 +2,12 @@
 
 use anchor_lang::prelude::*;
 
+mod state;
+mod instructions;
+mod errors;
+
+use instructions::*;
+
 declare_id!("2gt4rq13cxxjWBhfrdA8yXbhTwaqYAettE5rEuquqjgp");
 
 const INITIAL_PRIZE_POOL: u64 = 0;
@@ -23,14 +29,7 @@ pub mod blockrunners {
     }
 
     pub fn join_game(ctx: Context<JoinGame>) -> Result<()> {
-        let player_state = &mut ctx.accounts.player_state;
-
-        player_state.ciphers = 0;
-        player_state.cards = INITIAL_PLAYER_CARDS_AMOUNT; // Start with 1 card
-        player_state.position = 0;
-
-        msg!("Player joined the game.");
-        Ok(())
+        ctx.accounts.join_game(&ctx.bumps)
     }
 }
 
@@ -51,23 +50,6 @@ pub struct Initialize<'info> {
     pub system_program: Program<'info, System>,
 }
 
-#[derive(Accounts)]
-pub struct JoinGame<'info> {
-    #[account(mut)]
-    pub authority: Signer<'info>,
-
-    #[account(
-        init,
-        payer = authority,
-        space = 8 + 8 + 8 + 1, // discriminator + ciphers + cards + position
-        seeds = [b"player_state", authority.key().as_ref()],
-        bump
-    )]
-    pub player_state: Account<'info, PlayerState>,
-
-    pub system_program: Program<'info, System>,
-}
-
 #[account]
 pub struct GameState {
     /// The current prize pool amount in lamports
@@ -75,16 +57,4 @@ pub struct GameState {
 
     /// The length of the path players need to navigate (fixed at 20 for alpha)
     pub path_length: u8,
-}
-
-#[account]
-pub struct PlayerState {
-    /// Number of ciphers owned
-    pub ciphers: u64,
-
-    /// Number of cards 
-    pub cards: u64,
-
-    /// Current block number (fixed at 20 for alpha)
-    pub position: u8,
 }
